feat(manager): manager layout improvements
