+++ package.json
@@ -95,9 +95,9 @@
     "@types/redux": "^3.6.31",
     "@types/redux-devtools": "^3.0.34",
     "@types/redux-thunk": "^2.1.31",
+    "@types/sinon": "^1.16.32",
     "@types/webpack": "^2.0.0",
     "any-fs": "^0.1.0",
-    "@types/sinon": "^1.16.32",
     "babel-core": "^6.18.0",
     "babel-eslint": "^7.1.0",
     "babel-loader": "^6.2.7",
@@ -152,14 +152,15 @@
     "lodash": "^4.16.6",
     "makedeb": "0.0.4",
     "minimist": "^1.2.0",
-    "node-noop": "^1.0.0",
     "mocha": "^3.1.2",
+    "node-noop": "^1.0.0",
     "node-sass": "^3.10.1",
     "npm-install-webpack-plugin": "^4.0.4",
     "raw-loader": "^0.5.1",
     "rc-tree": "^1.3.9",
     "react": "^15.4.0",
     "react-addons-test-utils": "^15.4.0",
+    "react-dimensions": "^2.0.0-alpha1",
     "react-dom": "^15.4.0",
     "react-hot-loader": "^1.3.0",
     "react-icons": "^2.2.1",
